Microcleanup to combine the vars alltogether
git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@1078 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ scst/include/scst.h
@@ -1517,7 +1517,7 @@ struct scst_device {
 	unsigned short dev_double_ua_possible:1;
 
 	/* If set, dev is read only */
-	unsigned int rd_only:1;
+	unsigned short rd_only:1;
 
 	/**************************************************************/
